pyproject.toml
# Note that setup.py will install color_relief.sh script from the scripts directory

# python3 -m build --wheel
# python3 -m pip install .
# python3 -m pip install -e .

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ConfigEditor"
version = "0.1.11"
description = "Editor routines for text config and YML config files"
authors = [{ name = "corb" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PyYAML>=6.0.1",
    "PyQt6>=6.6.1",
]

[tool.setuptools]
# include the ConfigEditor directory as a package in the distribution.
packages = ["ConfigEditor"]

[tool.setuptools.package-dir]
# The ConfigEditor package is located in the root of the project
# directory (i.e., not inside a src/ or lib/ directory)
"" = "."
